+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Fixed
 
+- Fixed a layout issue on course glimpse
+  when organization title and course title are too long
 - Add missing language parameter in LTI requests issued by LTI consumer plugin
 
 ## [2.11.0] - 2022-01-04